+++ bin/commands/channels.rb
@@ -85,8 +85,10 @@ command 'channels' do |g|
   g.desc 'Lists all channels in a Slack team.'
   g.long_desc %( Lists all channels in a Slack team. )
   g.command 'list' do |c|
+    c.flag 'cursor', desc: "Paginate through collections of data by setting the cursor parameter to a next_cursor attribute returned by a previous request's response_metadata. Default value fetches the first 'page' of the collection. See pagination for more detail."
     c.flag 'exclude_archived', desc: 'Exclude archived channels from the list.'
     c.flag 'exclude_members', desc: 'Exclude the members collection from each channel.'
+    c.flag 'limit', desc: "The maximum number of items to return. Fewer than the requested number of items may be returned, even if the end of the users list hasn't been reached."
     c.action do |_global_options, options, _args|
       puts JSON.dump($client.channels_list(options))
     end